# Make observables 2-3 times faster to instantiate, and reduce their memory use by over half