@rniemeyer 2015-08-07
@SteveSanderson That looks great and makes perfect sense to me.