@SteveSanderson 2015-08-10
Looks like there were no objections so I've merged it.

Brian, I take your point about possibly being able to factor out more, and will look for an opportunity to do that as I work on computeds.