@harry8525 2015-08-10
Seems like we should have one precomputed prototype to extend here instead of possibly 3 (at least 2):

// Inherit from 'subscribable' 
27     if (ko.utils.canSetPrototype) { 
28         // 'subscribable' will come for free when we inherit from 'observable', so just set up the internal state needed for subscribables 
29         ko.subscribable['fn'].init(observable); 
30     } else { 
31         // 'subscribable' won't be on the prototype chain unless we put it there directly 
32         ko.subscribable.call(observable); 
33     } 
34 
 
35     // Inherit from 'observable' 
36     ko.utils.setPrototypeOfOrExtend(observable, observableFn); 
37 
 
38     if (ko.options['deferUpdates']) { 
39         ko.extenders['deferred'](observable, true); 
40     }